Get basic Python file read/write working
